drm: share --drm-list-modes across both KMS backends + report plane rotation
PrintDrmModes + ConnectorTypeName move from the drm_kms_egl backend into a
shared display/drm_mode_list.{h,cc} (libdrm-only, no GL/Vulkan/GBM), so
--drm-list-modes works identically on drm_kms_egl and drm_kms_vulkan; it was
wired only for EGL/software before. The listing now opens the device read-only
(safe alongside a live compositor) and adds a per-plane rotation-capability
section, since rotation is a plane property.
